backend/pyproject.toml
[tool.poetry]
name = "qa-grader"
version = "0.1.0"
description = "QA grading model"
authors = ["John <john@example.com>"]
readme = "README.md"
packages = [
  { include = "answer_classifier" },
  { include = "question_generator" },
  { include = "common" },
  { include = "dataparser" },
  { include = "api" },
  { include = "eda" },
  { include = "commands.py" }
]

[tool.poetry.dependencies]
python = "^3.10"
pytorch-lightning = "2.1.0"
transformers = "4.51.3"
datasets = "2.15.0"
torchmetrics = "1.6.1"
fire = "0.5.0"
hydra-core = "1.3.2"
mlflow = "2.19.0"
dvc = "3.35.0"
dvc-s3 = "3.2.0"
sentence-transformers = "4.1.0"
sentencepiece = ">=0.2.0"
openai = ">=1.0"
fastapi = ">=0.115.12"
torch = "2.2.0"
uvicorn = { extras = ["standard"], version = "*" }
numpy = "^1.24.0"
nltk = "=3.9.1"
bert-score = "=0.3.13"
seaborn = "0.13.2"
pandas = "=2.2.3"
matplotlib = "=3.10.1"
opentelemetry-api = "1.24.0"
opentelemetry-sdk = "1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pre-commit = "^3.0"
black = "^23.0"
isort = "^5.0"
flake8 = "^6.0"
mypy = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
